+++ src/lib/pages/account-details/components/tables/InstantiatedContractsTable.tsx
@@ -94,9 +94,10 @@ export const InstantiatedContractsTable = observer(
               isLoading={isLoading}
               emptyState={
                 !contracts ? (
-                  <ErrorFetching message="There is an error during fetching contracts." />
+                  <ErrorFetching dataName="contracts" />
                 ) : (
                   <EmptyState
+                    imageVariant="empty"
                     message="This account did not instantiate any contracts before."
                     withBorder
                   />