+++ src/lib/services/delegationService.ts
@@ -1,6 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { CELATONE_QUERY_KEYS, useBaseApiRoute } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useCelatoneApp,
+} from "lib/app-provider";
 import type { Addr } from "lib/types";
 
 import type {} from "./delegation";
@@ -8,11 +12,17 @@ import { getDelegationsByAddress } from "./delegation";
 
 export const useDelegationsByAddress = (address: Addr) => {
   const endpoint = useBaseApiRoute("accounts");
+  const {
+    chainConfig: {
+      extra: { disableDelegation },
+    },
+  } = useCelatoneApp();
 
   return useQuery(
     [CELATONE_QUERY_KEYS.DELEGATIONS_BY_ADDRESS, endpoint, address],
     () => getDelegationsByAddress(endpoint, address),
     {
+      enabled: !disableDelegation,
       refetchOnWindowFocus: false,
     }
   );