+++ src/lib/services/proposalService.ts
@@ -13,7 +13,10 @@ import {
   getProposalListCount,
   getProposalTypes,
   getRelatedProposalsByContractAddressPagination,
+  getRelatedProposalsByModuleIdPagination,
+  getRelatedProposalsCountByModuleId,
 } from "lib/query";
+import { createQueryFnWithTimeout } from "lib/query-utils";
 import type {
   ContractAddr,
   Option,
@@ -23,6 +26,7 @@ import type {
   Proposal,
   U,
   Token,
+  Nullish,
 } from "lib/types";
 import {
   deexponentify,
@@ -109,6 +113,80 @@ export const useProposalsByAddress = (
     ],
     async () => getProposalsByAddress(endpoint, address, limit, offset),
     { retry: 1, refetchOnWindowFocus: false }
+  );
+};
+
+export const useRelatedProposalsByModuleIdPagination = (
+  moduleId: Nullish<number>,
+  offset: number,
+  pageSize: number
+): UseQueryResult<Proposal[]> => {
+  const { indexerGraphClient } = useCelatoneApp();
+  const queryFn = useCallback(async () => {
+    if (!moduleId) throw new Error("Module id not found");
+    return indexerGraphClient
+      .request(getRelatedProposalsByModuleIdPagination, {
+        moduleId,
+        offset,
+        pageSize,
+      })
+      .then(({ module_proposals }) =>
+        module_proposals.map<Proposal>((proposal) => ({
+          proposalId: proposal.proposal_id,
+          title: proposal.proposal.title,
+          status: parseProposalStatus(proposal.proposal.status),
+          votingEndTime: parseDate(proposal.proposal.voting_end_time),
+          depositEndTime: parseDate(proposal.proposal.deposit_end_time),
+          resolvedHeight: proposal.proposal.resolved_height,
+          type: proposal.proposal.type as ProposalType,
+          proposer: proposal.proposal.account?.address as Addr,
+          isExpedited: Boolean(proposal.proposal.is_expedited),
+        }))
+      );
+  }, [indexerGraphClient, moduleId, offset, pageSize]);
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS_BY_MODULE_ID,
+      moduleId,
+      indexerGraphClient,
+      offset,
+      pageSize,
+    ],
+    queryFn,
+    {
+      retry: 1,
+      refetchOnWindowFocus: false,
+    }
+  );
+};
+
+export const useRelatedProposalsCountByModuleId = (
+  moduleId: Nullish<number>
+): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
+  const queryFn = useCallback(async () => {
+    if (!moduleId) throw new Error("Module id not found");
+    return indexerGraphClient
+      .request(getRelatedProposalsCountByModuleId, {
+        moduleId,
+      })
+      .then(
+        ({ module_proposals_aggregate }) =>
+          module_proposals_aggregate?.aggregate?.count
+      );
+  }, [indexerGraphClient, moduleId]);
+
+  return useQuery(
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS_COUNT_BY_MODULE_ID,
+      moduleId,
+      indexerGraphClient,
+    ],
+    createQueryFnWithTimeout(queryFn),
+    {
+      keepPreviousData: true,
+    }
   );
 };